Adding unit test project and unit tests for APIs
