+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "dotnet-test-explorer.testProjectPath": "**/*Tests.@(csproj|vbproj|fsproj)"
+}
