--- API/API.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="HotChocolate.AspNetCore" Version="11.3.6" />
-  </ItemGroup>
-
-</Project>
