+++ API/API.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26124.0
+MinimumVisualStudioVersion = 15.0.26124.0
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API", "API\API.csproj", "{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API.Tests", "API.Tests\API.Tests.csproj", "{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Debug|x64.Build.0 = Debug|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Debug|x86.Build.0 = Debug|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Release|x64.ActiveCfg = Release|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Release|x64.Build.0 = Release|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Release|x86.ActiveCfg = Release|Any CPU
+		{C41F3BE9-9BEE-4FDA-BD11-C7D3951E0B1C}.Release|x86.Build.0 = Release|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Debug|x64.Build.0 = Debug|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Debug|x86.Build.0 = Debug|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Release|x64.ActiveCfg = Release|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Release|x64.Build.0 = Release|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Release|x86.ActiveCfg = Release|Any CPU
+		{6FFFDD67-E7A7-425D-BA28-82E62D7B4ACE}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
