+++ API/API/Author.cs
@@ -1,0 +1,7 @@
+namespace API
+{
+    public class Author
+    {
+        public string Name { get; set; }
+    }
+}
