+++ API/API/netcoreapp3.1/API.runtimeconfig.dev.json
@@ -1,0 +1,8 @@
+{
+  "runtimeOptions": {
+    "additionalProbingPaths": [
+      "/Users/mattkellock/.dotnet/store/|arch|/|tfm|",
+      "/Users/mattkellock/.nuget/packages"
+    ]
+  }
+}
