--- API/Author.cs
@@ -1,7 +1,0 @@
-namespace API
-{
-    public class Author
-    {
-        public string Name { get; set; }
-    }
-}
