--- API/Book.cs
@@ -1,9 +1,0 @@
-namespace API
-{
-    public class Book
-    {
-        public string Title { get; set; }
-
-        public Author Author { get; set; }
-    }
-}
