--- API/Query.cs
@@ -1,15 +1,0 @@
-namespace API
-{
-    public class Query
-    {
-            public Book GetBook() =>
-            new Book
-            {
-                Title = "C# in depth.",
-                Author = new Author
-                {
-                    Name = "Jon Skeet"
-                }
-            };
-    }
-}
